fix main.rs
Fixed a bug where every win made by claiming the last slot was recognized as draw. This happend because the only condition for draw was that every slot is claimed. Now after checking for unclaimed slots it also checks if Player 1 [X] won, bc he is the one who claims the last slot. If all slots are claimed and Player 1 [X] didnt win a draw will happen.